+++ app/assets/javascripts/models/beta/story.js
@@ -63,6 +63,12 @@ export const isRelease = (storyType) => storyType === 'release';
 
 export const types = ['feature', 'bug', 'release', 'chore'];
 
+export const states = [
+  'unscheduled', 'unstarted', 'started',
+  'finished', 'delivered', 'accepted',
+  'rejected'
+];
+
 export const update = (story, projectId) => {
   const newStory = changeCase.snakeKeys(story);